+++ src/main/java/org/pahappa/systems/hms/constants/HospitalDepartment.java
@@ -1,4 +1,4 @@
-package constants;
+package org.pahappa.systems.hms.constants;
 
 public enum HospitalDepartment {
     // Clinical Departments
