+++ src/main/webapp/WEB-INF/includes/admin/manage_service_catalog.xhtml
@@ -18,7 +18,7 @@
                              value="#{serviceCatalogAdminBean.catalogItems}"
                              widgetVar="catalogTableWidget"
                              paginator="true" rows="10"
-                             emptyMessage="No services found in catalog."
+                             emptyMessage="No org.pahappa.systems.hms.navigation.services found in catalog."
                              scrollable="true" scrollHeight="300"
                              reflow="true">
